Remove layers on layer selector plugin is turn off
Each service-node in the layer selector tree now has a method
called hideAllLayers that will the esriMap to turn of the layers
associated with that service.

The plugin's hibernate() method will start at the root node of the
tree and walk down until it encounters a plugin with a hideAllLayers()
method, ie., a service node, at which point it calls that method.