Add methods for getState/setState on LayerManager
Create methods setServiceState and saveServiceState on the
individual service managers. setServiceState does what would
be expected: unpacks the state for each specific service, sets
the opacity and visible layers, and checks and expands the UI tree.
saveServiceState is used instead of a more typical "get", where
the service manager simply returns a value to its parent. This
is done so that the service manager can receive a state object for the
whole plugin, and choose how it wants to store its service data. This
is necessary because the different services need to store data of a
different schema, render into different structures in the UI tree, and
choose their own key.

Create methods for the LayerManager to walk the tree to service nodes
and apply the set+save algorithms.

Factor out complicated getLayerIds logic into a reusable function.

Rename layerIds to visibleLayerIds to be more explicit.

Rename public methods to be more consistent.