Only render the layer selector ext tree once
Fixes Github Issue #122

For some reason, subsequent renderings were causing the viewable that
the tree renders into to become smaller, probably due to packing more
dom elements into it.

Believed to fix Github Issue #112

It is unclear why this would fix this issue, but in tests the issue
appears to be resolved.