Add downloadUrl property to layer_selector plugin configuration.
You can now specify a downloadUrl property on layers defined in the showLayers group for AGS sources, and the layerIds group for WMS sources. This will add a button next to the layer name with the specified URL.

For AGS dynamic services without a showLayers group, add a new group called layerMetaData. This group closely resembles showLayers but is sparsely populated. Properties in this group will be merged with the original layers data source.